+++ packages/luna-stage/package.json
@@ -43,6 +43,7 @@
     "build": "rslib build"
   },
   "dependencies": {
+    "@dugyu/luna-core": "workspace:*",
     "clsx": "catalog:utils",
     "tailwind-merge": "catalog:tailwind"
   },
@@ -57,6 +58,7 @@
     "typescript": "catalog:ts"
   },
   "peerDependencies": {
+    "@lynx-js/web-core": ">=0.20.0",
     "motion": ">=12.23.16",
     "react": ">=18.0.0",
     "react-dom": ">=18.0.0"
