+++ packages/luna-stage/src/components/lynx-stage/index.ts
@@ -4,3 +4,9 @@
 
 export { LynxStage } from './lynx-stage';
 export type { LynxStageProps } from './lynx-stage';
+
+export { useLynxStage } from './use-lynx-stage';
+export type { UseLynxStageOptions, UseLynxStageResult } from './use-lynx-stage';
+
+export { LunaLynxStage } from './luna-lynx-stage';
+export type { LunaLynxStageProps } from './luna-lynx-stage';
