+++ packages/luna-stage/src/index.ts
@@ -13,3 +13,16 @@ export {
 // ─── Static phone mockup ──────────────────────────────────────────────────────
 export { Mockup, MockupContainer } from './components/mockup';
 export type { MockupProps } from './types';
+
+// ─── Lynx stage ────────────────────────────────────────────────────────────────
+export { LynxStage } from './components/lynx-stage';
+export type { LynxStageProps } from './components/lynx-stage';
+
+export { useLynxStage } from './components/lynx-stage';
+export type {
+  UseLynxStageOptions,
+  UseLynxStageResult,
+} from './components/lynx-stage';
+
+export { LunaLynxStage } from './components/lynx-stage';
+export type { LunaLynxStageProps } from './components/lynx-stage';
